requirements.txt
tensorflow==2.14
streamlit==1.28.0
Pillow==10.1.0